refactor(premium-analytics): move isSiteRegistered into site-sync, leave routing untouched
