+++ projects/packages/premium-analytics/packages/routing/package.json
@@ -7,10 +7,8 @@
 	"types": "src/index.ts",
 	"sideEffects": false,
 	"dependencies": {
-		"@automattic/jetpack-script-data": "workspace:*",
 		"@jetpack-premium-analytics/data": "workspace:*",
 		"@jetpack-premium-analytics/datetime": "workspace:*",
-		"@wordpress/data": "10.48.0",
 		"@wordpress/route": "0.12.0"
 	}
 }
