+++ projects/packages/premium-analytics/packages/routing/src/index.ts
@@ -6,4 +6,3 @@ export {
 
 export { deriveComparisonRange } from './search/comparison';
 export { useStagedSearch } from './hooks';
-export { isSiteRegistered } from './is-site-registered';
