+++ projects/packages/premium-analytics/packages/site-sync/src/index.ts
@@ -1,3 +1,4 @@
 export { useSyncStatus } from './hooks/use-sync-status';
+export { isSiteRegistered } from './is-site-registered';
 export { siteSyncStore, SITE_SYNC_STORE } from './store';
 export type { SyncStatus, SyncStatusApiResponse, UseSyncStatusReturn } from './types';
