+++ projects/packages/premium-analytics/packages/site-sync/src/store.ts
@@ -20,9 +20,7 @@ type SetMilestoneAction = {
 };
 
 const DEFAULT_STATE: State = {
-	// getScriptData() reads the inline script-data already present on the page,
-	// so this synchronous seed is correct on first load — before any poll runs
-	// and before the first route guard reads the store.
+	// Synchronous seed so the first route-guard read is correct before any poll runs.
 	milestone: getScriptData()?.premium_analytics?.initial_full_sync_finished ?? 0,
 };
 
@@ -35,9 +33,7 @@ const reducer = ( state: State = DEFAULT_STATE, action: SetMilestoneAction ): St
 
 const actions = {
 	/**
-	 * Set the initial-full-sync milestone. The sole writer (`useSyncStatus`)
-	 * advances it monotonically via `Math.max`, so in practice the value only
-	 * ever grows; the reducer itself stays a plain setter.
+	 * Set the initial-full-sync milestone.
 	 *
 	 * @param value - Milestone (unix ts).
 	 * @return Action object.
@@ -54,8 +50,7 @@ const selectors = {
 	 */
 	getMilestone: ( state: State ): number => state.milestone,
 	/**
-	 * Whether the initial full sync has finished — the fact the route guards
-	 * gate on.
+	 * Whether the initial full sync has finished.
 	 *
 	 * @param state - Store state.
 	 * @return Whether the initial full sync has finished.
@@ -66,10 +61,8 @@ const selectors = {
 /**
  * Site-sync data store.
  *
- * Holds the initial-full-sync milestone so synchronous route guards can answer
- * "has the initial sync finished?" without a full-page reload. Seeded from the
- * page-load script-data snapshot at registration, then advanced live by the
- * sync poll (see `useSyncStatus`).
+ * Holds the initial-full-sync milestone so synchronous route guards can read it.
+ * Seeded from script-data, then advanced live by the sync poll (see `useSyncStatus`).
  */
 export const siteSyncStore = createReduxStore( SITE_SYNC_STORE, {
 	reducer,
@@ -77,8 +70,6 @@ export const siteSyncStore = createReduxStore( SITE_SYNC_STORE, {
 	selectors,
 } );
 
-// Registered as a side-effect of importing this module: route guards that
-// `select( siteSyncStore )` import it transitively, so the store always exists
-// before any `beforeLoad` runs. `package.json` lists this file in `sideEffects`
-// so the registration is not tree-shaken away.
+// Registered on import so guards can `select()` it before any `beforeLoad`.
+// Listed in `package.json` `sideEffects` so the registration survives tree-shaking.
 register( siteSyncStore );
