+++ projects/packages/premium-analytics/routes/connect/package.json
@@ -8,7 +8,7 @@
 	"dependencies": {
 		"@automattic/jetpack-connection": "workspace:*",
 		"@automattic/jetpack-script-data": "workspace:*",
-		"@jetpack-premium-analytics/routing": "workspace:*",
+		"@jetpack-premium-analytics/site-sync": "workspace:*",
 		"@wordpress/components": "35.0.0",
 		"@wordpress/element": "8.0.0",
 		"@wordpress/i18n": "^6.9.0",
