+++ projects/packages/premium-analytics/routes/connect/route.tsx
@@ -1,15 +1,12 @@
 /**
  * External dependencies
  */
-import { isSiteRegistered } from '@jetpack-premium-analytics/routing';
+import { isSiteRegistered } from '@jetpack-premium-analytics/site-sync';
 import { redirect } from '@wordpress/route';
 
 /**
- * Route guard for /connect.
- * If the site is already connected, redirect to the dashboard.
- *
- * Registration is read live (see `isSiteRegistered`), so a client-side
- * registration is reflected without a full-page reload.
+ * Route guard for /connect: redirect to the dashboard if the site is already
+ * registered.
  */
 export const route = {
 	beforeLoad: () => {
