+++ projects/packages/premium-analytics/routes/dashboard/route.tsx
@@ -1,8 +1,7 @@
 /**
  * External dependencies
  */
-import { isSiteRegistered } from '@jetpack-premium-analytics/routing';
-import { siteSyncStore } from '@jetpack-premium-analytics/site-sync';
+import { isSiteRegistered, siteSyncStore } from '@jetpack-premium-analytics/site-sync';
 import { select } from '@wordpress/data';
 import { redirect } from '@wordpress/route';
 
@@ -10,9 +9,6 @@ import { redirect } from '@wordpress/route';
  * Route guard for the dashboard.
  * - Not connected → /connect
  * - Connected but sync pending → /syncing
- *
- * Both facts are read from live stores, so the guard routes correctly after a
- * client-side registration or sync completion without a full-page reload.
  */
 export const route = {
 	beforeLoad: () => {
