+++ projects/packages/premium-analytics/routes/syncing/package.json
@@ -6,7 +6,6 @@
 		"page": "jetpack-premium-analytics"
 	},
 	"dependencies": {
-		"@jetpack-premium-analytics/routing": "workspace:*",
 		"@jetpack-premium-analytics/site-sync": "workspace:*",
 		"@wordpress/components": "35.0.0",
 		"@wordpress/data": "10.48.0",
